add re-parent subject tests, for PUT and PATCH
